Testing

+++ code/game/turfs/open/lava.dm
@@ -2,13 +2,14 @@
 
 /turf/open/lava
 	name = "lava"
+	desc = "Thick viscous bubbles swell and burst with sprays of molten rock. What was once locked away in the bowels of the earth now thrashes and gurgles on the surface, glowing with unimaginable heat. If you touch it, you will die."
 	icon_state = "lava"
 	icon = 'icons/turf/roguefloor.dmi'
 	gender = PLURAL //"That's some lava."
 	baseturfs = /turf/open/lava //lava all the way down
 	slowdown = 2
 
-	light_range = 4
+	light_outer_range =  4
 	light_power = 0.75
 	light_color = LIGHT_COLOR_LAVA
 	bullet_bounce_sound = 'sound/blank.ogg'
@@ -21,6 +22,9 @@
 	canSmoothWith = list(/turf/closed, /turf/open/floor/rogue/volcanic, /turf/open/floor/rogue/dirt, /turf/open/floor/rogue/dirt/road,/turf/open/floor/rogue/naturalstone)
 	neighborlay_override = "lavedge"
 
+/turf/open/lava/nosmooth
+	smooth = SMOOTH_FALSE
+
 /turf/open/lava/Initialize()
 	. = ..()
 	dir = pick(GLOB.cardinals)
@@ -28,6 +32,8 @@
 /turf/open/lava/cardinal_smooth(adjacencies)
 	roguesmooth(adjacencies)
 
+/turf/open/lava/can_traverse_safely(atom/movable/traveler)
+	return FALSE
 
 /turf/open/lava/ex_act(severity, target)
 	contents_explosion(severity, target)
@@ -45,9 +51,6 @@
 /turf/open/lava/MakeDry(wet_setting = TURF_WET_WATER)
 	return
 
-/turf/open/lava/airless
-	initial_gas_mix = AIRLESS_ATMOS
-
 /turf/open/lava/Entered(atom/movable/AM)
 	if(!AM.throwing)
 		if(burn_stuff(AM))
@@ -56,70 +59,54 @@
 			playsound(src, 'sound/misc/lava_death.ogg', 100, FALSE)
 //			addomen("lava")
 
-/turf/open/lava/Exited(atom/movable/Obj, atom/newloc)
-	. = ..()
-	if(!Obj.throwing)
-		if(isliving(Obj))
-			var/mob/living/L = Obj
-			if(!islava(newloc) && !L.on_fire)
-				L.update_fire()
-
-/turf/open/lava/hitby(atom/movable/AM, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum, d_type = "blunt")
+/turf/open/lava/hitby(atom/movable/AM, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum, damage_flag = "blunt")
 	if(burn_stuff(AM))
 		START_PROCESSING(SSobj, src)
 		playsound(src, 'sound/misc/lava_death.ogg', 100, FALSE)
 
 /turf/open/lava/process()
-	if(!burn_stuff())
+	if(!burn_stuff()) // try to burn everything in our contents, stop once nothing left can burn
 		STOP_PROCESSING(SSobj, src)
-
-/turf/open/lava/rcd_vals(mob/user, obj/item/construction/rcd/the_rcd)
-	switch(the_rcd.mode)
-		if(RCD_FLOORWALL)
-			return list("mode" = RCD_FLOORWALL, "delay" = 0, "cost" = 3)
-	return FALSE
-
-/turf/open/lava/rcd_act(mob/user, obj/item/construction/rcd/the_rcd, passed_mode)
-	switch(passed_mode)
-		if(RCD_FLOORWALL)
-			to_chat(user, span_notice("I build a floor."))
-			PlaceOnTop(/turf/open/floor/plating, flags = CHANGETURF_INHERIT_AIR)
-			return TRUE
-	return FALSE
-
-/turf/open/lava/singularity_act()
-	return
-
-/turf/open/lava/singularity_pull(S, current_size)
-	return
 
 /turf/open/lava/get_smooth_underlay_icon(mutable_appearance/underlay_appearance, turf/asking_turf, adjacency_dir)
 	underlay_appearance.icon = 'icons/turf/floors.dmi'
 	underlay_appearance.icon_state = "basalt"
 	return TRUE
 
-/turf/open/lava/GetHeatCapacity()
-	. = 700000
-
-/turf/open/lava/GetTemperature()
-	. = 5000
-
-/turf/open/lava/TakeTemperature(temp)
-
-
 /turf/open/lava/proc/is_safe()
 	//if anything matching this typecache is found in the lava, we don't burn things
-	var/static/list/lava_safeties_typecache = typecacheof(list(/obj/structure/lattice/catwalk, /obj/structure/stone_tile))
+	var/static/list/lava_safeties_typecache = typecacheof(list(/obj/structure/stone_tile))
 	var/list/found_safeties = typecache_filter_list(contents, lava_safeties_typecache)
 	for(var/obj/structure/stone_tile/S in found_safeties)
 		if(S.fallen)
 			LAZYREMOVE(found_safeties, S)
 	return LAZYLEN(found_safeties)
 
+/turf/open/lava/can_traverse_safely(atom/movable/traveler)
+	return ..() && !will_burn(traveler) // can traverse safely if you won't burn in it
+
+/turf/open/lava/proc/will_burn(atom/movable/thing)
+	if(isobj(thing))
+		var/obj/O = thing
+		if((O.resistance_flags & (LAVA_PROOF|INDESTRUCTIBLE)) || O.throwing)
+			return FALSE
+		return TRUE
+	else if (isliving(thing))
+		var/mob/living/L = thing
+		if(L.movement_type & FLYING)
+			return FALSE //YOU'RE FLYING OVER IT
+		if("lava" in L.weather_immunities) // just flat-out immune. is this even used in RT?
+			return FALSE
+		var/buckle_check = L.buckling
+		if(!buckle_check)
+			buckle_check = L.buckled
+		if(buckle_check && !will_burn(buckle_check))
+			return FALSE // buckled to something lavaproof
+		return TRUE
+	return FALSE // no handling for this type burning, obj or living only
 
 /turf/open/lava/proc/burn_stuff(AM)
-	. = 0
-
+	. = FALSE
 	if(is_safe())
 		return FALSE
 
@@ -127,12 +114,12 @@
 	if (AM)
 		thing_to_check = list(AM)
 	for(var/thing in thing_to_check)
+		if(!will_burn(thing))
+			continue
 		if(isobj(thing))
 			var/obj/O = thing
-			if((O.resistance_flags & (LAVA_PROOF|INDESTRUCTIBLE)) || O.throwing)
-				continue
-			. = 1
-			if((O.resistance_flags & (ON_FIRE)))
+			. = TRUE
+			if((O.resistance_flags & (ON_FIRE))) // already on fire, don't bother. why do we do this exactly...? is this bad copypasta?
 				continue
 			if(!(O.resistance_flags & FLAMMABLE))
 				O.resistance_flags |= FLAMMABLE //Even fireproof things burn up in lava
@@ -143,45 +130,47 @@
 			qdel(O)
 
 		else if (isliving(thing))
-			. = 1
+			. = TRUE
 			var/mob/living/L = thing
-			if(L.movement_type & FLYING)
-				continue	//YOU'RE FLYING OVER IT
-			var/buckle_check = L.buckling
-			if(!buckle_check)
-				buckle_check = L.buckled
-			if(isobj(buckle_check))
-				var/obj/O = buckle_check
-				if(O.resistance_flags & LAVA_PROOF)
-					continue
-			else if(isliving(buckle_check))
-				var/mob/living/live = buckle_check
-				if("lava" in live.weather_immunities)
-					continue
-
-			if(!L.on_fire)
-				L.update_fire()
 
 			if(iscarbon(L))
 				var/mob/living/carbon/C = L
 				var/obj/item/clothing/S = C.get_item_by_slot(SLOT_ARMOR)
 				var/obj/item/clothing/H = C.get_item_by_slot(SLOT_HEAD)
 
+				// we still catch fire if wearing lavaproof armor, but we don't get dusted when dead
+				// is this really the intended behaviour, or was it just badly coded? idk
 				if(S && H && S.clothing_flags & LAVAPROTECT && H.clothing_flags & LAVAPROTECT)
-					return
-
-				if(C.health <= 0)
-					C.dust(drop_items = TRUE)
+					continue
 
 			if("lava" in L.weather_immunities)
 				continue
 
-//			L.adjustFireLoss(50)
-			if(L) //mobs turning into object corpses could get deleted here.
+			if(L)
+				L.adjustFireLoss(100)
 				L.adjust_fire_stacks(100)
-				L.IgniteMob()
-				if(L.health <= 0)
-					L.dust(drop_items = TRUE)
+				L.ignite_mob()
+
+/turf/open/lava/onbite(mob/user)
+	if(isliving(user))
+		var/mob/living/L = user
+		if(L.stat != CONSCIOUS)
+			return
+		if(iscarbon(user))
+			var/mob/living/carbon/C = user
+			if(C.is_mouth_covered())
+				return
+		playsound(user, pick('sound/foley/waterwash (1).ogg','sound/foley/waterwash (2).ogg'), 100, FALSE)
+		user.visible_message(span_info("[user] starts to drink from [src]."))
+		if(do_after(L, 25, target = src))
+			var/mob/living/carbon/C = user
+			to_chat(C, span_userdanger("OH SWEET PSYDON, WHY DID I THINK THIS WAS A GOOD IDEA???"))
+			C.flash_fullscreen("redflash3")
+			C.emote("agony", forced = TRUE)
+			C.adjust_fire_stacks(500) //you deserve this.
+			C.ignite_mob()
+			C.adjustFireLoss(1000) //you, literally, deserve this.
+
 /turf/open/lava/smooth
 	name = "lava"
 	baseturfs = /turf/open/lava/smooth
@@ -191,19 +180,15 @@
 	canSmoothWith = list(/turf/open/lava/smooth)
 
 /turf/open/lava/smooth/lava_land_surface
-	initial_gas_mix = LAVALAND_DEFAULT_ATMOS
-	planetary_atmos = TRUE
+
 	baseturfs = /turf/open/lava/smooth/lava_land_surface
 
 /turf/open/lava/smooth/airless
-	initial_gas_mix = AIRLESS_ATMOS
 
 /turf/open/lava/acid
 	name = "acid"
-	desc = "Caustic and horrid, it bubbles incessantly."
 	icon_state = "acid"
-	slowdown = 8 //caustic cove edit
-	light_range = 4
+	light_outer_range =  4
 	light_power = 1
 	light_color = "#56ff0d"
 
@@ -221,8 +206,10 @@
 			var/obj/O = thing
 			if((O.resistance_flags & (ACID_PROOF|INDESTRUCTIBLE)) || O.throwing)
 				continue
+			O.obj_integrity -= O.max_integrity * 0.1
+			if(O.obj_integrity <= 0)
+				qdel(O)	
 			. = 1
-			qdel(O)
 
 		else if (isliving(thing))
 			. = 1
@@ -240,39 +227,31 @@
 				var/mob/living/live = buckle_check
 				if("lava" in live.weather_immunities)
 					continue
+			for(var/obj/item/clothing/C in L.contents)
+				if(C.resistance_flags & (ACID_PROOF|INDESTRUCTIBLE))
+					continue
+				C.obj_integrity -= C.max_integrity * 0.1
+				if(C.obj_integrity <= 0)
+					to_chat(L, span_danger("Your [C.name] is destroyed by the acid!"))
+					qdel(C)	
 
-//			if(iscarbon(L))
-//				var/mob/living/carbon/C = L
-//				var/obj/item/clothing/S = C.get_item_by_slot(SLOT_ARMOR)
-//				var/obj/item/clothing/H = C.get_item_by_slot(SLOT_HEAD)
+			L.adjustFireLoss(100)
+			to_chat(L, span_userdanger("THE ACID BURNS!"))
 
-//				if(S && H && S.clothing_flags & LAVAPROTECT && H.clothing_flags & LAVAPROTECT)
-//					return
-				//make this acid
-/***
-				var/shouldupdate = FALSE
-				var/lethality = prob(95)
-				for(var/obj/item/bodypart/B in C.bodyparts)
-					if(!B.skeletonized && B.is_organic_limb())
-						B.skeletonize(lethality)
-						shouldupdate = TRUE
-				if(!lethality)
-					ADD_TRAIT(C, TRAIT_NOLIMBDISABLE, "[type]")
-				if(shouldupdate)
-					if(ishuman(C))
-						var/mob/living/carbon/human/H = C
-						qdel(H.underwear)
-					C.unequip_everything()
-					C.update_body()
-//				C.dust(drop_items = TRUE)
-				continue
-
-//			if("lava" in L.weather_immunities)
-//				continue
-***/
-
-			L.adjustFireLoss(10) //would have higher, but I wanted to give prey their fanservice
-			playsound(src, 'modular_causticcove/sound/misc/bigmelt.ogg', 100, FALSE)
-			if(L.health <= 0) //melt away once dead
-				L.dust(drop_items = TRUE)
-				playsound(src, 'modular_causticcove/sound/misc/deathdigest.ogg', 100, FALSE) //caustic cove edit end
+/turf/open/lava/acid/onbite(mob/user)
+	if(isliving(user))
+		var/mob/living/L = user
+		if(L.stat != CONSCIOUS)
+			return
+		if(iscarbon(user))
+			var/mob/living/carbon/C = user
+			if(C.is_mouth_covered())
+				return
+		playsound(user, pick('sound/foley/waterwash (1).ogg','sound/foley/waterwash (2).ogg'), 100, FALSE)
+		user.visible_message(span_info("[user] starts to drink from [src]."))
+		if(do_after(L, 25, target = src))
+			var/mob/living/carbon/C = user
+			to_chat(C, span_userdanger("OH SWEET PSYDON, WHY DID I THINK THIS WAS A GOOD IDEA???"))
+			C.flash_fullscreen("redflash3")
+			C.emote("agony", forced = TRUE)
+			C.gib() //YOU. LITERALLY. DESERVE THIS.
